+++ lib/hoodoo/client.rb
@@ -17,6 +17,7 @@ require 'hoodoo/client/endpoint/endpoint'
 require 'hoodoo/client/endpoint/endpoints/http_based'
 require 'hoodoo/client/endpoint/endpoints/http'
 require 'hoodoo/client/endpoint/endpoints/amqp'
+require 'hoodoo/client/endpoint/endpoints/new_relic_traced_amqp'
 require 'hoodoo/client/endpoint/endpoints/not_found'
 require 'hoodoo/client/endpoint/endpoints/auto_session'
